requirements.txt
# Supabase client
supabase==2.3.4

# Football data scraping
soccerdata==1.5.0

# Data manipulation (soccerdata dependency, pinned for stability)
pandas==2.1.4
numpy==1.26.4

# HTTP & HTML parsing (soccerdata dependency)
requests==2.31.0
lxml==5.1.0
beautifulsoup4==4.12.3

# Timezone handling
pytz==2024.1

# Environment variables
python-dotenv==1.0.1
